fix(storage): satisfy durable merge gates
